processor-py/requirements.txt
pandas>=2.2
numpy>=1.26
Flask>=3.0
jinja2>=3.1
pydantic>=2.7
pyyaml>=6.0
